+++ packages/provider-utils/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ai-sdk/provider-utils
 
+## 3.1.0-beta.1
+
+### Patch Changes
+
+- cbb1d35: Update for provider-util changeset after change in PR #8588
+
+## 3.1.0-beta.0
+
+### Minor Changes
+
+- 78928cb: release: start 5.1 beta
+
+### Patch Changes
+
+- Updated dependencies [78928cb]
+  - @ai-sdk/provider@2.1.0-beta.0
+
 ## 3.0.9
 
 ### Patch Changes